fix: move loading.tsx to root, replace generateMetadata with metadata
