+++ src/app/auth/sign-in/page.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from "react";
 import { SignInForm } from "@/components/auth/SignInForm";
-import AuthLoading from "../loading";
+import AuthLoading from "../../loading";
 import { createMetadata } from "@/lib/metadata";
 
 export const metadata = createMetadata({
